Add Export history component to export
